expense-tracker-streamlit/requirements.txt
streamlit>=1.28.0
gspread>=6.0.0
google-auth>=2.23.0
bcrypt>=4.0.0
